post 1차 수정:잘못된 접근에 대한  조건문 추가

+++ src/routes/posts.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 const pool = require('../config/db');
 const path = require('path'); //추후 삭제가능성 있음
-const { isLoggedIn, isNotLoggedIn} = require('./middlewares');
+const { isLoggedIn } = require('./middlewares');
 
 /*
     router끼리의 순서 중요
@@ -10,19 +10,20 @@ const { isLoggedIn, isNotLoggedIn} = require('./middlewares');
 */
 
 // 게시글 생성 페이지 불러오기  get /post/create
-router.get('/create', (req, res, next) => {
+router.get('/create', isLoggedIn, (req, res, next) => {
     res.sendFile(path.resolve('views/testCreatePost.html')); // 추후 변경
 })
 
 // 게시글 생성하기  post /post/create
 // 이미지 삽입 --> 추후 구현
-router.post('/create', async (req,res,next) => {
-    const { user_id, title, content, category } = req.body;
+router.post('/create', isLoggedIn, async (req,res,next) => {
+    const {title, content, category } = req.body;
+    const userId = req.user.id;
     try {
         // 게시글 DB에 추가하기
         await pool.execute(
             "INSERT INTO post(user_id, title, content, category) VALUES(?,?,?,?)",
-            [user_id, title, content, category]
+            [userId, title, content, category]
         );
         return res.redirect('/'); // 생성된 게시글페이지로 이동 --> 추후 변경
     } catch (error) {
@@ -41,18 +42,18 @@ router.get('/:category', async (req, res, next) => {
 
     // 카테고리가 존해하지 않을 경우
     if (!categoryList.includes(category)) {
+        // :id 라우터로 이동
         return next();
     }
     // 카테고리가 존재하는 경우
     try {
         // DB에서 해당 카테고리 게시글 목록 불러오기
         const [dbPosts] = await pool.execute(
-            `SELECT p.id, u.profile_img, u.nickname, p.title, p.content, p.view_count, p.comment_count, p.category, p.created_date 
+            `SELECT p.id, u.profile_img, unickname, p.title, p.content, p.view_count, p.comment_count, p.category, p.created_date 
             FROM post p LEFT JOIN user u ON p.user_id = u.id
             WHERE p.category = ?;`,
             [category]
         );
-        console.log(dbPosts);
         return res.status(201).json({dbPosts}); // 추후 변경
     } catch (error) {
         console.log(error);
@@ -64,7 +65,7 @@ router.get('/:category', async (req, res, next) => {
 })
 
 // 세부 게시글 페이지 불러오기 get /post/:id
-router.get('/:id',async (req, res, next) => {
+router.get('/:id', isLoggedIn, async (req, res, next) => {
     const postId = req.params.id;
     try {
         // DB에서 해당 id의 게시글과 댓글 목록 가져오기
@@ -87,9 +88,25 @@ router.get('/:id',async (req, res, next) => {
 })
 
 // 세부 게시글 수정 페이지 불러오기  get /post/:id/edit
-router.get('/:id/edit', async (req, res, next) => {
+router.get('/:id/edit', isLoggedIn, async (req, res, next) => {
     const postId = req.params.id;
+    const userId = req.user.id;
     try {
+        // DB에서 post를 쓴 유저 id값 불러오기
+        const [dbPostUser] = await pool.execute(
+            `SELECT user_id FROM post where id=?;`,
+            [postId]
+        );
+        // post_id가 잘못된 값인 경우
+        if (Array.isArray(dbPostUser) && dbPostUser.length == 0) {
+            console.log('잘못된 post_id 값');
+            return res.redirect('/');
+        }
+        // post를 쓴 userId 와 세션 유저 id가 다른 경우
+        if (dbPostUser[0].user_id !== userId){
+            console.log('접근 오류')
+            return res.redirect('/');
+        }
         // DB에서 수정할 게시글 가져오기
         const [dbPost] = await pool.execute(
             `SELECT u.nickname, u.profile_img, p.title, p.content, p.category 
@@ -97,10 +114,6 @@ router.get('/:id/edit', async (req, res, next) => {
             where p.id=?;`,
             [postId]
         );
-        // post_id가 잘못된 값인 경우
-        if (Array.isArray(dbPost) && dbPost.length == 0) {
-            return res.redirect('/');
-        }
         return res.status(201).json({dbPost}); // 추후 변경
     } catch (error) {
         console.log(error);
@@ -109,10 +122,26 @@ router.get('/:id/edit', async (req, res, next) => {
 })
 
 // 세부 게시글 수정하기 patch /post/:id
-router.patch('/:id', async (req,res,next) => {
+router.patch('/:id', isLoggedIn, async (req,res,next) => {
     const postId = req.params.id;
+    const userId = req.user.id;
     const {title, content, category} = req.body;
     try {
+        // DB에서 post를 쓴 유저 id값 불러오기
+        const [dbPostUser] = await pool.execute(
+            `SELECT user_id FROM post where id=?;`,
+            [postId]
+        );
+        // post_id가 잘못된 값인 경우
+        if (Array.isArray(dbPostUser) && dbPostUser.length == 0) {
+            console.log('잘못된 post_id 값');
+            return res.redirect('/');
+        }
+        // post를 쓴 userId 와 세션 유저 id가 다른 경우
+        if (dbPostUser[0].user_id !== userId){
+            console.log('접근 오류')
+            return res.redirect('/');
+        }
         // DB에 해당 id 게시글 수정하기
         await pool.execute(
             "UPDATE post SET title=?, content=?, category=? where id=?",
@@ -125,9 +154,25 @@ router.patch('/:id', async (req,res,next) => {
     }
 });
 // 세부 게시글 삭제하기 delete /post/:id
-router.delete('/:id', async (req,res,next) => {
+router.delete('/:id', isLoggedIn, async (req,res,next) => {
     const postId = req.params.id;
+    const userId = req.user.id;
     try {
+        // DB에서 post를 쓴 유저 id값 불러오기
+        const [dbPostUser] = await pool.execute(
+            `SELECT user_id FROM post where id=?;`,
+            [postId]
+        );
+        // post_id가 잘못된 값인 경우
+        if (Array.isArray(dbPostUser) && dbPostUser.length == 0) {
+            console.log('잘못된 post_id 값');
+            return res.redirect('/');
+        }
+        // post를 쓴 userId 와 세션 유저 id가 다른 경우
+        if (dbPostUser[0].user_id !== userId){
+            console.log('접근 오류')
+            return res.redirect('/');
+        }
         // DB에 해당 id 게시글 삭제
         await pool.execute(
             "DELETE FROM post WHERE id=?;",
